Refactor:CurationTabContent 컴포넌트 아토믹 디자인으로 리팩토링
