+++ app/(pages)/(main)/curation/page.tsx
@@ -1,8 +1,9 @@
 import getScrappedCuration from "@feature/curation/queries/getScrappedCuration";
-import CurationTabContent from "@feature/curation/components/CurationTabContent/CurationTabContent";
+import CurationTabContent from "@feature/curation/components/CurationInfo/organisms/CurationTabContent";
 import getMyCuration from "@feature/curation/queries/getMyCuration";
 import PageTopBar from "@common/components/ui/topBar/PageTopBar/PageTopBar";
 
+//Page
 export default async function CurationPage() {
   const myCuration = await getMyCuration();
   const scrappedCuration = await getScrappedCuration();
