+++ app/(pages)/(main)/curation/popular/page.tsx
@@ -1,5 +1,5 @@
 import ArrowBackTopBar from "@common/components/ui/topBar/ArrowBackTopBar/ArrowBackTopBar";
-import CurationCardLight from "@feature/curation/components/CurationCardLight/CurationCardLight";
+import CurationInfoCardLight from "@feature/curation/components/CurationInfo/molecules/CurationInfoCardLight";
 import GetRandomCuration from "@feature/curation/queries/getRandomCuration";
 
 export default async function CurationPopularPage() {
@@ -10,7 +10,7 @@ export default async function CurationPopularPage() {
       <div className="text-black headline1 px-[2rem]">마포구 인기 큐레이션</div>
       <section className="mt-[2.4rem] px-[2rem] pb-[12rem]">
         {randomCuration.map((curation: any) => (
-          <CurationCardLight
+          <CurationInfoCardLight
             key={curation.author + curation.id}
             {...curation}
             className="mb-[2rem] w-full"
