+++ app/feature/curation/components/CurationInfo/molecules/MyCurationCount.tsx
@@ -1,0 +1,12 @@
+interface MyCurationCountProps {
+  count: number;
+}
+
+//Molecule
+export default function MyCurationCount({ count }: MyCurationCountProps) {
+  return (
+    <div className="flex body1 text-text-gray-8 items-center">
+      총 <p className="text-black">&nbsp;{count}</p>개
+    </div>
+  );
+}
