+++ app/feature/curation/components/CurationInfo/stories/CurationInfoCardDark.stories.tsx
@@ -1,10 +1,10 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import CurationCardLight from "./CurationCardLight";
+import CurationInfoCardDark from "../molecules/CurationInfoCardDark";
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
 const meta = {
-  title: "Feature/Curation/CurationCardLight/CurationCardLight",
-  component: CurationCardLight,
+  title: "Feature/Curation/CurationScrapped/CurationScrapped",
+  component: CurationInfoCardDark,
   parameters: {
     // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/configure/story-layout
     layout: "centered",
@@ -13,7 +13,7 @@ const meta = {
   tags: ["autodocs"],
   // More on argTypes: https://storybook.js.org/docs/api/argtypes
   argTypes: {},
-} satisfies Meta<typeof CurationCardLight>;
+} satisfies Meta<typeof CurationInfoCardDark>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
@@ -23,11 +23,12 @@ export const ScrappedFalse: Story = {
   args: {
     id: 0,
     image: [
-      "https://cdn.pixabay.com/photo/2023/10/24/08/24/sailboats-8337698_1280.jpg",
+      "https://cdn.pixabay.com/photo/2016/12/03/15/44/fireworks-1880045_1280.jpg",
     ],
     author: "김현민",
     title: "크리스마스에 즐기기 좋은 마포구 데이트 코스",
     keyword: ["연인과의 데이트", "크리스마스"],
+    isScraped: true,
   },
 };
 
@@ -35,10 +36,11 @@ export const ScrappedTrue: Story = {
   args: {
     id: 0,
     image: [
-      "https://cdn.pixabay.com/photo/2023/10/24/08/24/sailboats-8337698_1280.jpg",
+      "https://cdn.pixabay.com/photo/2016/12/03/15/44/fireworks-1880045_1280.jpg",
     ],
     author: "김현민",
     title: "크리스마스에 즐기기 좋은 마포구 데이트 코스",
     keyword: ["연인과의 데이트", "크리스마스"],
+    isScraped: true,
   },
 };
